+++ src/main/java/firok/mtim/MoreTinkersMaterials.java
@@ -28,7 +28,7 @@ public class MoreTinkersMaterials
 
 	public static final String MOD_ID = "mtim";
 	public static final String MOD_NAME = "MoreTinkersMaterials";
-	public static final String VERSION = "1.12.2-0.1.6";
+	public static final String VERSION = "1.12.2-0.1.7";
 
 
 	{